Add threaded handler.
NOTE: airbrake.submit.submit_payload is not yet guaranteed thread-safe.

Also, moved report generation from the handlers and into the tween, since I
don't think it's strictly safe to be passing dicts around threads, even for
read-only operations. (I believe that it is actually safe to perform reads on
dicts between threads in CPython, but that this is an implementation detail and
that thread safety is not guaranteed in other implementations).